# apps/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/builds/{build_id}": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドを取得",
        "operationId": "get_build",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ビルド情報",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/builds/{build_id}/logs": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドの進捗ログ（増分取得）",
        "description": "render / compare ジョブが追記した進捗ログを `after` カーソルで増分取得する。テナントのメンバーであること。進行中は UI がポーリングする。",
        "operationId": "get_build_logs",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "この id より後の行だけを返す（省略時は 0 = 先頭から）。",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "進捗ログの行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildLogsResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/builds/{build_id}/storybook/": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "アップロード済み Storybook を開く（index.html）",
        "description": "`mode = storybook` のビルドにアップロードされた Storybook を対話的に配信する エントリポイント。認可はビルド詳細と同じ（テナントメンバー、または `read:build` の PAT）。バンドル未アップロードや screenshots モードは 404。",
        "operationId": "get_storybook_index",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Storybook の index.html",
            "content": {
              "text/html": {}
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/ci/builds/{build_id}": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "ビルドの状態を取得する（CI のポーリング用）",
        "description": "`read:build` スコープが必要。`status` が終端になるまで CI がポーリングする。",
        "operationId": "get_build_status",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ビルドの状態とカウント",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildLogsResponse": {
        "type": "object",
        "description": "ビルド進捗ログの増分取得レスポンス。",
        "required": [
          "entries",
          "last_id"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuildLogEntry"
            },
            "description": "`after` より後の行（id 昇順）。"
          },
          "last_id": {
            "type": "integer",
            "format": "int64",
            "description": "クライアントが次のポーリングで `after` に渡す値。\n行が無ければリクエストされた `after` が据え置かれる。"
          }
        }
      },
      "BuildMode": {
        "type": "string",
        "description": "ビルドの入力形式。\n\n- [`BuildMode::Screenshots`]: CI が自前で撮った PNG をアップロードする（従来モード）\n- [`BuildMode::Storybook`]: CI がビルド済み Storybook の zip をアップロードし、\n  サーバー側（`RenderBuildJob`）がヘッドレス Chromium で全ストーリーを撮る",
        "enum": [
          "screenshots",
          "storybook"
        ]
      },
      "BuildResponse": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "number",
          "branch",
          "commit_sha",
          "status",
          "mode",
          "storybook_uploaded",
          "total_count",
          "changed_count",
          "added_count",
          "removed_count",
          "unchanged_count",
          "created_at"
        ],
        "properties": {
          "added_count": {
            "type": "integer",
            "format": "int32"
          },
          "approved_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "approved_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "baseline_commit_sha": {
            "type": [
              "string",
              "null"
            ],
            "description": "このビルドが比較する baseline のコミット SHA。\n\n将来の CLI が「どのコミットとの差分か」を知り、\n撮り直しの必要なストーリーを自分で絞り込めるように公開する。\nビルド作成（`POST /v1/ci/builds`）のレスポンスでのみ埋まり、\nそれ以外の経路（一覧・状態ポーリング等）では常に `None`。\nbaseline が無い、または昇格元ビルドが削除済みなら `None`。"
          },
          "baseline_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "比較に使った baseline（未確定なら null）。"
          },
          "branch": {
            "type": "string"
          },
          "changed_count": {
            "type": "integer",
            "format": "int32"
          },
          "commit_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "commit_sha": {
            "type": "string"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/BuildMode",
            "description": "入力形式（`screenshots` = CI がアップロード / `storybook` = サーバーがレンダリング）。"
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "description": "プロジェクト内で連番のビルド番号。"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "pull_request_number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "removed_count": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/BuildStatus"
          },
          "storybook_uploaded": {
            "type": "boolean",
            "description": "storybook モードでバンドルがアップロード済みか。"
          },
          "total_count": {
            "type": "integer",
            "format": "int32"
          },
          "unchanged_count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BuildStatus": {
        "type": "string",
        "description": "ビルドのライフサイクル。遷移は `service::builds::transition` に集約する。\n\n```text\nscreenshots モード:\npending ──finalize──▶ processing ──▶ passed | changes_detected | failed\n\nstorybook モード:\npending ──finalize──▶ rendering ──▶ processing ──▶ passed | changes_detected | failed\n                         └──────────────────────▶ failed\n\nどちらも: changes_detected ──▶ approved | rejected / passed ──▶ approved\n```",
        "enum": [
          "pending",
          "rendering",
          "processing",
          "passed",
          "changes_detected",
          "failed",
          "approved",
          "rejected"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT"
      }
    }
  },
...
}